add confirmation

+++ editor.js
@@ -26,13 +26,15 @@ document.getElementById("service0").focus();
 var selectedType = "service";
 var selectedServiceId=0;
 function save() {
-	len = Object.keys(secrets).length;
-	secrets = {};
-	for (var i = 0; i<len; i++) {
-		secrets[document.getElementById("service"+i).value] = {"secret":(document.getElementById("secret"+i).value)} 
+	if (confirm("Save changes and exit?")===true) {
+		len = Object.keys(secrets).length;
+		secrets = {};
+		for (var i = 0; i<len; i++) {
+			secrets[document.getElementById("service"+i).value] = {"secret":(document.getElementById("secret"+i).value)} 
+		}
+		localStorage.setItem("secrets", JSON.stringify(secrets));
+		window.location.href = "index.html";
 	}
-	localStorage.setItem("secrets", JSON.stringify(secrets));
-	window.location.href = "index.html";
 }
 function exitwithoutsaving() {
 	if (confirm("Exit without saving?")===true) {
